(PC-16709) refactor(cookies): add types and enums files
