+++ src/features/cookies/helpers/getAcceptedCookieConsent.ts
@@ -1,4 +1,4 @@
-import { CookieNameEnum } from 'features/cookies/CookiesPolicy'
+import { CookieNameEnum } from 'features/cookies/enums'
 import { getCookiesChoice as getCookiesConsentV2 } from 'features/cookies/helpers/useCookies'
 
 export const getAcceptedCookieConsent = async (cookie: CookieNameEnum): Promise<boolean | null> => {
